+++ saleor/static/dashboard-next/navigation/components/MenuItemDialog/MenuItemDialog.tsx
@@ -4,6 +4,7 @@ import DialogActions from "@material-ui/core/DialogActions";
 import DialogContent from "@material-ui/core/DialogContent";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import TextField from "@material-ui/core/TextField";
+import Typography from "@material-ui/core/Typography";
 import isUrl from "is-url";
 import React from "react";
 
@@ -12,11 +13,15 @@ import ConfirmButton, {
   ConfirmButtonTransitionState
 } from "@saleor/components/ConfirmButton";
 import FormSpacer from "@saleor/components/FormSpacer";
-import { SearchCategories_categories_edges_node } from "../../../containers/SearchCategories/types/SearchCategories";
-import { SearchCollections_collections_edges_node } from "../../../containers/SearchCollections/types/SearchCollections";
-import { SearchPages_pages_edges_node } from "../../../containers/SearchPages/types/SearchPages";
-import i18n from "../../../i18n";
-import { getMenuItemByValue, IMenu } from "../../../utils/menu";
+import { SearchCategories_categories_edges_node } from "@saleor/containers/SearchCategories/types/SearchCategories";
+import { SearchCollections_collections_edges_node } from "@saleor/containers/SearchCollections/types/SearchCollections";
+import { SearchPages_pages_edges_node } from "@saleor/containers/SearchPages/types/SearchPages";
+import useModalDialogErrors from "@saleor/hooks/useModalDialogErrors";
+import useStateFromProps from "@saleor/hooks/useStateFromProps";
+import i18n from "@saleor/i18n";
+import { UserError } from "@saleor/types";
+import { getErrors, getFieldError } from "@saleor/utils/errors";
+import { getMenuItemByValue, IMenu } from "@saleor/utils/menu";
 
 export type MenuItemType = "category" | "collection" | "link" | "page";
 export interface MenuItemData {
@@ -31,6 +36,7 @@ export interface MenuItemDialogFormData extends MenuItemData {
 export interface MenuItemDialogProps {
   confirmButtonState: ConfirmButtonTransitionState;
   disabled: boolean;
+  errors: UserError[];
   initial?: MenuItemDialogFormData;
   initialDisplayValue?: string;
   loading: boolean;
@@ -68,6 +74,7 @@ function getDisplayValue(menu: IMenu, value: string): string {
 const MenuItemDialog: React.StatelessComponent<MenuItemDialogProps> = ({
   confirmButtonState,
   disabled,
+  errors: apiErrors,
   initial,
   initialDisplayValue,
   loading,
@@ -79,10 +86,11 @@ const MenuItemDialog: React.StatelessComponent<MenuItemDialogProps> = ({
   collections,
   pages
 }) => {
+  const errors = useModalDialogErrors(apiErrors, open);
   const [displayValue, setDisplayValue] = React.useState(
     initialDisplayValue || ""
   );
-  const [data, setData] = React.useState<MenuItemDialogFormData>(
+  const [data, setData] = useStateFromProps<MenuItemDialogFormData>(
     initial || defaultInitial
   );
   const [url, setUrl] = React.useState<string>(undefined);
@@ -97,6 +105,8 @@ const MenuItemDialog: React.StatelessComponent<MenuItemDialogProps> = ({
     setDisplayValue(initialDisplayValue);
     setUrl(undefined);
   }, [open]);
+
+  const mutationErrors = getErrors(errors);
 
   let options: IMenu = [];
 
@@ -192,6 +202,10 @@ const MenuItemDialog: React.StatelessComponent<MenuItemDialogProps> = ({
 
   const handleSubmit = () => onSubmit(data);
 
+  const idError = ["category", "collection", "page", "url"]
+    .map(field => getFieldError(errors, field))
+    .reduce((acc, err) => acc || err);
+
   return (
     <Dialog
       onClose={onClose}
@@ -203,9 +217,13 @@ const MenuItemDialog: React.StatelessComponent<MenuItemDialogProps> = ({
       }}
     >
       <DialogTitle>
-        {i18n.t("Add Item", {
-          context: "create new menu item"
-        })}
+        {!!initial
+          ? i18n.t("Edit Item", {
+              context: "edit menu item"
+            })
+          : i18n.t("Add Item", {
+              context: "create new menu item"
+            })}
       </DialogTitle>
       <DialogContent style={{ overflowY: "visible" }}>
         <TextField
@@ -220,22 +238,31 @@ const MenuItemDialog: React.StatelessComponent<MenuItemDialogProps> = ({
             }))
           }
           name="name"
-          helperText=""
+          error={!!getFieldError(errors, "name")}
+          helperText={getFieldError(errors, "name")}
         />
         <FormSpacer />
         <AutocompleteSelectMenu
           disabled={disabled}
           onChange={handleSelectChange}
           name="id"
-          helperText=""
           label={i18n.t("Link")}
           displayValue={displayValue}
           loading={loading}
-          error={false}
           options={options}
+          error={!!idError}
+          helperText={idError}
           placeholder={i18n.t("Start typing to begin search...")}
           onInputChange={handleQueryChange}
         />
+        {mutationErrors.length > 0 && (
+          <>
+            <FormSpacer />
+            {mutationErrors.map(err => (
+              <Typography color="error">{err}</Typography>
+            ))}
+          </>
+        )}
       </DialogContent>
       <DialogActions>
         <Button onClick={onClose}>